fix(report-card): revert mobile vote-btns relocation; match rating pill style to confidence
Moving vote-btns into .signal-strip on mobile caused them to wrap to a
fresh row anyway (signal icons + votes > mobile width). Reverts back
to the original layout where votes render inline with the confidence
pill and rating badge on mobile (via the existing .right flex-row).

Also matches .card .rating styling to .card .confidence-pill so the
pair reads as a single group -- shared 3px radius, 0.72rem size, 700
weight, 0.04em letter-spacing, 3px 9px padding.

Refs #165
