+++ game-stats.html
@@ -15,7 +15,7 @@
 <link rel="stylesheet" href="css/shared/cards.css?v=16c748ed">
 <link rel="stylesheet" href="css/app/game-header.css?v=594fa872">
 <link rel="stylesheet" href="css/app/configs.css?v=90d5bed5">
-<link rel="stylesheet" href="css/app/reports.css?v=ad76fb62">
+<link rel="stylesheet" href="css/app/reports.css?v=a9280434">
 <link rel="stylesheet" href="css/app/modals.css?v=37997d79">
 <link rel="stylesheet" href="css/app/home.css?v=95acc50a">
 <style>
